+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -193,6 +193,11 @@ New checks
 
   Checks that different identifiers are typographically unambiguous.
 
+- New :doc:`bsl-implicit-conversions-forbidden
+  <clang-tidy/checks/bsl-implicit-conversions-forbidden>` check.
+
+  Warns of any implicit conversion
+
 - New :doc:`bsl-lambda-implicit-capture
   <clang-tidy/checks/bsl-lambda-implicit-capture>` check.